requirements.txt
pydantic>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
openai>=1.0.0
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
